+++ example/lib/main.dart
@@ -48,6 +48,10 @@ class _MyAppState extends State<MyApp> {
     _clevertapPlugin.setCleverTapDisplayUnitsLoadedHandler(onDisplayUnitsLoaded);
     _clevertapPlugin.setCleverTapInAppNotificationButtonClickedHandler(inAppNotificationButtonClicked);
     _clevertapPlugin.setCleverTapInboxNotificationButtonClickedHandler(inboxNotificationButtonClicked);
+    _clevertapPlugin.setCleverTapFeatureFlagUpdatedHandler(featureFlagsUpdated);
+    _clevertapPlugin.setCleverTapProductConfigInitializedHandler(productConfigInitialized);
+    _clevertapPlugin.setCleverTapProductConfigFetchedHandler(productConfigFetched);
+    _clevertapPlugin.setCleverTapProductConfigActivatedHandler(productConfigActivated);
   }
 
   void inAppNotificationDismissed(Map<String,dynamic> map){
@@ -132,6 +136,26 @@ class _MyAppState extends State<MyApp> {
       List displayUnits = await CleverTapPlugin.getAllDisplayUnits();
       print("Display Units = "+ displayUnits.toString());
     });
+  }
+
+  void featureFlagsUpdated(){
+    print("Feature Flags Updated");
+    this.setState(() async {
+      bool booleanVar = await CleverTapPlugin.getFeatureFlag("FFKey", false);
+      print("Feature flag = " + booleanVar.toString());
+    });
+  }
+
+  void productConfigInitialized(){
+    print("Product Config Initialized");
+  }
+
+  void productConfigFetched(){
+    print("Product Config Fetched");
+  }
+
+  void productConfigActivated(){
+    print("Product Config Activated");
   }
 
   @override
@@ -350,6 +374,18 @@ class _MyAppState extends State<MyApp> {
                 RaisedButton(
                     onPressed: () => pushInboxNotificationViewedEventForId(),
                     child: Text('pushInboxNotificationViewedEventForId')
+                ),
+                RaisedButton(
+                    onPressed: () => fetch(),
+                    child: Text('Fetch')
+                ),
+                RaisedButton(
+                    onPressed: () => activate(),
+                    child: Text('Activate')
+                ),
+                RaisedButton(
+                    onPressed: () => fetchAndActivate(),
+                    child: Text('Fetch And Activate')
                 ),
               ],
             ),
@@ -871,6 +907,18 @@ class _MyAppState extends State<MyApp> {
   void getAdUnits() async{
     List displayUnits = await CleverTapPlugin.getAllDisplayUnits();
     print("Display Units = "+ displayUnits.toString());
-  }  
+  }
+
+  void fetch(){
+    CleverTapPlugin.fetch();
+  }
+
+  void activate(){
+    CleverTapPlugin.activate();
+  }
+
+  void fetchAndActivate(){
+    CleverTapPlugin.fetchAndActivate();
+  }
 
 }